Ongoing campaign type (#33)
* Calendar shortcode

* Set up ongoing sign up shortcode

* Ongoing: Setup create subscription

* Ongoing Calendar and Selector upgrades

* Split subscription management to separate files

* upgrade small calendar

* Upgrade subscription management

* Bulk update and delete prayer times

* enable ongoing short codes

* Fix inline control structure.

* Better function and var names
