+++ magic-links/ongoing/ongoing.css
@@ -1,0 +1,190 @@
+.cp-wrapper {
+    max-width:1000px;
+    margin: 0 auto;
+    padding: 2em 1em;
+    border-radius: 10px;
+    font-size: 1em;
+    min-width: 320px;
+}
+.cp-center {
+    text-align: center;
+}
+.cp-wrapper strong, b {
+    font-weight: bold;
+}
+.cp-wrapper {
+    font-size: 1rem;
+}
+.cp-wrapper h3 {
+    font-size: 1.5rem;
+}
+.cp-wrapper h2 {
+    font-size: 1.7rem;
+}
+.cp-wrapper h1, .cp-wrapper h2, .cp-wrapper h3 {
+    color: black;
+}
+.cp-wrapper label {
+    font-size: 1rem;
+    margin-bottom: .5rem;
+}
+
+.cp-wrapper p {
+    margin: 10px;
+}
+/**
+ * Calendar section
+ */
+.cp-calendar-wrapper {
+    background-color: #f8f9fad1;
+    border-radius: 10px;
+    padding: 1em
+}
+.cp-calendar-wrapper .day-cell {
+    text-align: center;
+    flex-grow: 0;
+}
+.cp-calendar-wrapper .disabled-calendar-day {
+    width:40px;
+    height:40px;
+    vertical-align: top;
+    padding-top:10px;
+    color: grey;
+}
+.cp-calendar-wrapper .calendar {
+    display: flex;
+    flex-wrap: wrap;
+    width: 300px;
+    margin: auto
+}
+.cp-calendar-wrapper .month-title {
+    text-align: left;
+    margin-bottom: 5px;
+}
+.cp-calendar-wrapper .calendar-month {
+    display: inline-block;
+    vertical-align: top;
+}
+.cp-calendar-wrapper .week-day {
+    height: 20px;
+    width:40px;
+    color:black;
+    font-size:12px;
+    font-weight:550;
+    margin-bottom:5px;
+}
+.cp-calendar-wrapper .goto-month-container {
+    display: flex;
+    align-items: center;
+    width:40px;
+}
+.cp-calendar-wrapper .cp-goto-month {
+    width:40px;
+}
+
+/**
+ * Sign up section
+ */
+.cp-wrapper #email, #email-later {
+    display:none;
+}
+.cp-wrapper .day-in-select-calendar {
+    color: black;
+    display: inline-block;
+    height: 40px;
+    width: 40px;
+    line-height: 0;
+    vertical-align: middle;
+    text-align: center;
+    padding-top: 18px;
+}
+.cp-wrapper .selected-day {
+    color: white;
+    border-radius: 50%;
+    border: 2px solid;
+}
+
+.cp-wrapper .success-confirmation-section {
+    display: none;
+    margin-top: 20px;
+}
+
+.cp-wrapper .cp-close-button {
+    top: .5rem;
+    font-size: 1em;
+    line-height: 1;
+    display: block;
+    cursor:pointer;
+    padding: 5px;
+}
+.cp-wrapper .cp-close-button img {
+    filter: invert(100%);
+    height: 15px;
+    width: 15px;
+    vertical-align: bottom;
+}
+.cp-wrapper select {
+    font-size: 1rem;
+    line-height: 1rem;
+    color: black;
+    border: 1px solid black ;
+    display: block;
+    min-width: 250px;
+    max-width: 400px;
+    background: white;
+    margin: auto;
+    min-height: 40px;
+    padding:0.5em;
+}
+.cp-wrapper input {
+    font-size: 1rem;
+    line-height: 1rem;
+    color: black;
+    border: 1px solid black;
+}
+.cp-wrapper .cp-input {
+    min-width: 250px;
+    max-width: 400px;
+    margin: auto;
+    padding:0.5em;
+    min-height: 40px;
+}.cp-wrapper.loading-content h2, .cp-wrapper.loading-content p {
+    background-color: #ededed;
+    border-radius: 100px;
+    min-width: 100px;
+    min-height: 20px;
+    margin-bottom: 5px;
+}
+
+.cp-wrapper button {
+    color: #fefefe;
+    font-size: 1rem;
+    border-radius: 5px;
+    border: 1px solid transparent;
+    font-weight: normal;
+    padding: .85em 1em;
+}
+.cp-wrapper button:hover {
+    background-color: transparent;
+}
+.cp-wrapper button[disabled] {
+    opacity: .25;
+    cursor: not-allowed;
+}
+
+.cp-wrapper .form-error {
+    display: none;
+    color: red;
+}
+
+.cp-wrapper ul {
+    list-style-type: none;
+    padding: 0;
+}
+
+.cp-wrapper .remove-prayer-time-button, .cp-wrapper .remove-daily-time-button {
+    background-color: transparent;
+    color: red;
+    /* border: black 1px solid; */
+    padding: 0 5px;
+}
