+++ magic-links/subscription-management/subscription-management.css
@@ -1,0 +1,392 @@
+body {
+    background-color: white;
+}
+#content {
+    max-width:100%;
+    min-height: 300px;
+    margin-bottom: 200px;
+}
+#title {
+    font-size:1.7rem;
+    font-weight: 100;
+}
+#wrapper {
+    max-width:1000px;
+    margin:0 auto 300px auto;
+    padding: .5em;
+    background-color: white;
+}
+/* Chrome, Safari, Edge, Opera */
+input::-webkit-outer-spin-button,
+input::-webkit-inner-spin-button {
+    -webkit-appearance: none;
+    margin: 0;
+}
+/* Firefox */
+input[type=number] {
+    -moz-appearance: textfield;
+}
+select::-ms-expand {
+    display: none;
+}
+
+/* size specific style section */
+@media screen and (max-width: 991px) {
+    /* start of large tablet styles */
+
+}
+@media screen and (max-width: 767px) {
+    /* start of medium tablet styles */
+
+}
+@media screen and (max-width: 479px) {
+    /* start of phone styles */
+    body {
+        background-color: white;
+    }
+}
+.remove-my-prayer-time {
+    color: red;
+    cursor:pointer;
+}
+.day-selector {
+    cursor: pointer;
+}
+#calendar-content .day-cell {
+    max-width: 25%;
+    float:left;
+    text-align: center;
+    border: 1px solid lightgray;
+    border-top: none;
+    border-left: 0.5px;
+    border-right: 0.5px;
+}
+
+.new-day-number {
+    padding-top: 18px;
+    padding-bottom: 1px;
+}
+
+.progress-bar {
+    height:10px;
+    background: dodgerblue;
+    width:0;
+}
+.progress-bar-container {
+    border: 1px solid #bfbfbf;
+    margin: 0 auto 12px auto;
+    width: 75%;
+}
+
+.remove-selection {
+    /*float: right;*/
+    color: white;
+    cursor:pointer;
+}
+
+.day-extra {
+    text-align: start;
+    padding: 5px 5px;
+}
+
+.day-cell {
+    /*flex-basis: 14%;*/
+    text-align: center;
+    flex-grow: 0;
+
+}
+.disabled-calendar-day {
+    width:40px;
+    height:40px;
+    vertical-align: top;
+    padding-top:10px;
+    color: grey;
+}
+.calendar {
+    display: flex;
+    flex-wrap: wrap;
+    width: 300px;
+    margin: auto
+}
+.month-title {
+    text-align: center;
+    margin-bottom: 0;
+}
+.week-day {
+    height: 20px;
+    width:40px;
+    color: grey;
+}
+#calendar-content h3 {
+    margin-bottom: 0;
+}
+
+#list-day-times td, th {
+    padding: 2px 6px
+}
+
+.day-in-select-calendar {
+    color: black;
+    display: inline-block;
+    height: 40px;
+    width: 40px;
+    line-height: 0;
+    vertical-align: middle;
+    text-align: center;
+    padding-top: 18px;
+}
+.selected-day {
+    background-color: dodgerblue;
+    color: white;
+    border-radius: 50%;
+    border: 2px solid;
+}
+
+#modal-calendar.small-view {
+    display: flex;
+    width: 250px;
+    flex-wrap: wrap;
+    margin: auto;
+}
+
+.small-view .calendar {
+    width: 180px;
+}
+.small-view .month-title {
+    width: 50px;
+    padding: 0 10px;
+    overflow: hidden;
+    flex-grow: 1;
+}
+.small-view .calendar .week-day {
+    height: 20px;
+    width: 25px;
+}
+.small-view .calendar .day-in-select-calendar {
+    height: 25px;
+    width: 25px;
+}
+.small-view .calendar .disabled-calendar-day {
+    width: 25px;
+    height: 25px;
+    padding-top: 5px;
+    font-size: 11px;
+}
+
+.small-view .day-in-select-calendar {
+    height: 25px;
+    width: 25px;
+    padding-top: 11px;
+    font-size:8px;
+}
+
+.cp-wrapper .confirm-view {
+    display: none;
+}
+.cp-wrapper .close-button {
+    font-size: 2em;
+    color:#8a8a8a;
+    background-color: transparent;
+    padding:0;
+}
+.cp-wrapper .success-confirmation-section {
+    display: none;
+}
+.deleted-time {
+    color: grey;
+    text-decoration: line-through;
+}
+.timezone-label {
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+    margin: 12px 0 24px 0;
+    text-align: right;
+}
+.timezone-label svg {
+    margin-right: 3px;
+}
+.calendar-title {
+    margin: 32px auto 0 auto;
+    display: flex;
+    justify-content: space-between;
+    align-items: baseline;
+    width: 84%;
+}
+.calendar-title h2{
+    color: dodgerblue;
+    font-weight: bold;
+}
+
+/* todo replace 'new_' with '' and unify css new styles with old ones */
+.new_calendar {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: left;
+    width: 92%;
+    margin: auto;
+    flex-grow: 1;
+}
+
+.new_day_cell {
+    width: calc(100% / 7);
+    height: auto;
+    text-align: center;
+    cursor: pointer;
+}
+
+.new-day-number:hover {
+    background-color: #eee;
+}
+.calendar-month {
+    flex-grow: 1;
+}
+.calendar-month .goto-month-container {
+    display: flex;
+    padding-top:150px;
+    width:40px;
+}
+.calendar-month .cp-goto-month {
+    width:40px;
+    height:40px;
+}
+
+.new_weekday {
+    width: calc(100% / 7);
+    height: auto;
+    text-align: center;
+    font-weight: bold;
+    margin: 12px auto 4px auto;
+    border-bottom: 2px solid black;
+    border-radius: 0%;
+    justify-content: left;
+}
+
+.prayer-commitment {
+    background-color: #57d449;
+    border-radius: 5%;
+    color: white;
+    font-size: 12px;
+    margin: 2px auto 2px auto;
+    text-align: right;
+}
+
+.prayer-commitment-text {
+    margin-right: 4px;
+}
+.prayer-commitment-tiny {
+    background-color: #57d449;
+    border-radius: 100%;
+    text-indent: 200%;
+    white-space: nowrap;
+    overflow: hidden;
+    width: 12px;
+    height: 12px;
+    vertical-align: top;
+    margin: -10px auto 0 auto;
+}
+
+#mobile-commitments-container {
+    display: none;
+}
+
+.mobile-commitments {
+    display: flex;
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+    vertical-align: middle;
+}
+
+.mobile-commitments-date {
+    color: #0a0a0a;
+    border-right: 2px solid darkgray;
+    text-align: center;
+    width: 25%;
+}
+
+.mc-day {
+    color: darkgray;
+}
+
+.mc-prayer-commitment-description {
+    color: white;
+    font-size: 12px;
+    margin: 0px auto 0px auto;
+    text-align: right;
+    width: 75%;
+    max-height: 100%;
+    padding: 6px;
+    vertical-align: middle;
+}
+
+.mc-prayer-commitment-text {
+    width: 100%;
+    background-color: #57d449;
+    color: white;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    border-radius: 5px;
+    margin: 0 0 5px 0;
+}
+
+.mc-title {
+    display: none;
+    margin: 50px 0 15px 10px;
+    font-size: 32px;
+    font-weight: bold;
+    text-align: center;
+}
+
+.mc-description-duration {
+    font-size: 125%;
+    font-weight: 400;
+    margin-left: 15px;
+}
+
+.mc-description-time {
+    font-size: 180%;
+    margin-right: 10px;
+}
+
+
+/**
+Danger zone
+ */
+.danger-zone{
+    margin: 0 8px 0 8px;
+    display: flex;
+    justify-content: space-between;
+}
+
+.chevron img{
+    vertical-align: middle;
+    width: 20px;
+    cursor: pointer;
+}
+
+.danger-zone-content {
+    display: flex;
+    justify-content: space-between;
+    margin: 24px 10% 0 5%;
+}
+.collapsed {
+    display: none;
+}
+
+.toggle_up {
+    -moz-transform: scale(-1, -1);
+    -o-transform: scale(-1, -1);
+    -webkit-transform: scale(-1, -1);
+    transform: scale(-1, -1);
+}
+.cp-wrapper button {
+    color: #fefefe;
+    font-size: 1rem;
+    border-radius: 5px;
+    border: 1px solid transparent;
+    font-weight: normal;
+    padding: .85em 1em;
+    background-color: dodgerblue;
+}
